worker-manager/requirements.txt
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
google-auth==2.18.1
idna==3.4
kubernetes==26.1.0
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.0.3
python-dateutil==2.8.2
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==1.5.1
